Debug convergence

+++ finemo/hitcaller.py
@@ -173,7 +173,7 @@ def fit_batch(cwms_t, contribs, sequences, coef_init, clip_mask,
             # tbatch.set_postfix(max_gap=gap.max().item(), mean_gap=gap.mean().item())
             num_nonzero = c_a.count_nonzero((1,2),) ####
             tbatch.set_postfix(max_gap=gap.max().item(), mean_gap=gap.mean().item(), max_ll=ll.max().item(), mean_ll=ll.mean().item(), 
-                               max_nonzero=num_nonzero.max().item(), mean_nonzero=num_nonzero.mean().item()) ####
+                               max_nonzero=num_nonzero.max().item(), mean_nonzero=num_nonzero.float().mean().item()) ####
 
 
             if torch.all(gap <= convergence_tol).item():
